Splitting and joining if/unless clauses in eruby
Switch between:

    <%= foo if bar? %>

And:

    <% if bar? %>
      <%= foo %>
    <% end %>

This is probably not very useful, since in erb, you'd usually have pure
text instead of an erb block in the body of the clause, but it's here
mostly for completion's sake -- to match the behaviour in ruby files.
Plus, it's fairly simple to do, at least to this extent, so it's a good
start.