First pass C# support by piggybacking on C support
I do not know whether C support will work 100% for C# but it works for
what I want (function parameters) and it is better than not supporting
it at all. If there does turn out to be divergence between C and C# then
C# should probably not piggyback on C support anymore, but for now I
don't see any harm in it.