Fix subtle ruby options bug
If a round-bracket function was detected away from the cursor, the
LocateFunction call will find it and only later will the splitting
function check for that and bail out. Instead, there might be a hash
around the cursor and a function call within it that's away from it.

Moving the check inside the LocateFunction call fixes this issue,
letting it correctly return 0, and keep the splitting callback going.