refine sj#elixir#SplitPipe()
* Handle multiple calls on one line (by doing nothing)
* Handle lines ending with a comment

The function call parser is not able to determine the end of calls that
don’t use parentheses. It returns a function_end of -1 in those cases.
However we can parse out the comment to deal with it.

Though this would break down if we have a ‘#’ inside a string, like
this:

foo "bar#baz" # bla

I think this is probably good enough for now.